Manage /etc/resolv.conf
Add functionality to control /etc/resolv.conf.
To control DNS systemd-resolved needs /etc/resolv.conf set as a symlink
to point to either the dynamic stub resolver file, the static stub
resolver file or a generated file containing all known DNS servers
(which bypasses systemd-resolved).
With this commit the role can do this for you and allows you to point
the /etc/resolv.conf to a file of your choosing.